back-end/requirements.txt
Flask
werkzeug
pyjwt